Update Immediately
In memory of Kevin, Falkor, and Gemma: the best three dogs a man could have. I'm sorry I couldn't do better for you and I'm sorry that the world was a confusing place. It should have never become so complicated. All Humans have to do is to learn how to share.